+++ src/components/Dropdown/DropdownContentInner.tsx
@@ -8,14 +8,15 @@ import { Rect, getContentBoxStyle, ContentBoxStyle } from './dropdownHelper'
 type Props = {
   triggerRect: Rect
   children: React.ReactNode
+  className: string
 }
 
-export const DropdownContentInner: FC<Props> = ({ triggerRect, children }) => {
+export const DropdownContentInner: FC<Props> = ({ triggerRect, children, className }) => {
   const theme = useTheme()
   const [isMounted, setIsMounted] = useState(false)
   const [contentBox, setContentBox] = useState<ContentBoxStyle>({
-    top: 'auto',
-    left: 'auto',
+    top: '0',
+    left: '0',
     maxHeight: '',
   })
   const wrapperRef = useRef<HTMLDivElement>(null)
@@ -50,7 +51,7 @@ export const DropdownContentInner: FC<Props> = ({ triggerRect, children }) => {
     <Wrapper
       ref={wrapperRef}
       contentBox={contentBox}
-      className={isMounted ? 'active' : ''}
+      className={`${className} ${isMounted ? 'active' : ''}`}
       themes={theme}
     >
       {children}